+++ bindings/python/rumi/__init__.py
@@ -1,5 +1,6 @@
 from importlib.metadata import version
 
+from ._checksums import get_checksum_verification, set_checksum_verification
 from ._frames import Frame, FrameTable, frames
 from ._info import Metadata, info
 from ._read import RumiArray, read, read_many
@@ -9,5 +10,6 @@ from ._write import write
 __version__ = version("rumi-eo")
 
 __all__ = ["Frame", "FrameTable", "Metadata", "RumiArray", "frames",
-           "get_num_threads", "info", "read", "read_many", "set_num_threads",
-           "write", "__version__"]
+           "get_checksum_verification", "get_num_threads", "info", "read",
+           "read_many", "set_checksum_verification", "set_num_threads", "write",
+           "__version__"]